Join the database network
(We have cookies)

The `digests_wsgi_1` container tries to connect to localhost, but this isn't the host EC2 instance: it's the container itself. To correctly connect to the `postgres` container, we need to join its network and use the `postgres` host. Databases provided by containers are on a separate network, fostering their isolation from the application.